index.js - made fakeData method that sends over fake data for testing state. script2.js - added listener for state which communicates with front end to display state. style.css - changed dimensions of info bars.
attempted to make d3 work to. no. avail. first i tried changing the fill underneath, but the animation starts spasming out the fill. then i thought of simplifying it s.t. the line color changes in different states. i tried to make dataEye1 an array of objects with keys 'eye1' and 'state' so i could access the state and change things from there. BUT THE ARRAY KEPT ON POPULATING WITH ZEROS AND I COULDN'T FIGURE OUT WHY BECAUSE ISN'T THERE ONLY ONE PLACE WE'RE MODIFYING THE ARRAY?? that was upsetting. currently i have state displayed on one of the info bars on the right. hopefully i'll figure out how to do the line coloring thing and will be able to build up these features!